#4: Add warning if user tries to delete customer in list-customers

+++ src/main/webapp/WEB-INF/views/list-customers.jsp
@@ -43,7 +43,7 @@
                     <td>
                         <a href="${updateLink}">Update</a>
                         |
-                        <a href="${deleteLink}">Delete</a>
+                        <a href="${deleteLink}" onclick="if (!(confirm('Are you sure you want to delete this customer?'))) return false">Delete</a>
                     </td>
                 </tr>
             </c:forEach>
